rename parallel functions library
