+++ freegs4e/test_parallel.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from scipy.special import ellipe, ellipk
 
-from . import parallel_funcs
+from . import parallel
 
 # All tests are written under the assumption that the parallel functions are implemented as wrappers, and should give
 # EXACTLY equivalent results to the serial versions
@@ -15,12 +15,12 @@ def make_test_array():
 
 def test_set_threads():
 
-    num_threads = parallel_funcs.get_num_threads()
-    parallel_funcs.set_num_threads(num_threads + 2)
+    num_threads = parallel.get_num_threads()
+    parallel.set_num_threads(num_threads + 2)
 
-    assert parallel_funcs.get_num_threads() == (num_threads + 2)
+    assert parallel.get_num_threads() == (num_threads + 2)
 
-    parallel_funcs.set_num_threads(num_threads)
+    parallel.set_num_threads(num_threads)
 
 
 def test_take(make_test_array):
@@ -28,23 +28,23 @@ def test_take(make_test_array):
     idcs = np.random.randint(0, 100, size=(20, 3))
     reference = np.take(make_test_array, idcs)
 
-    orig_num_threads = parallel_funcs.get_num_threads()
-    parallel_funcs.set_num_threads(1)
+    orig_num_threads = parallel.get_num_threads()
+    parallel.set_num_threads(1)
     assert np.all(
-        reference == parallel_funcs.threaded_take(make_test_array, idcs)
+        reference == parallel.threaded_take(make_test_array, idcs)
     ), "Thin wrapper fail (failed at 1 thread)"
 
-    parallel_funcs.set_num_threads(2)
+    parallel.set_num_threads(2)
     assert np.all(
-        reference == parallel_funcs.threaded_take(make_test_array, idcs)
+        reference == parallel.threaded_take(make_test_array, idcs)
     ), "Failed on 2 threads"
 
-    parallel_funcs.set_num_threads(61)
+    parallel.set_num_threads(61)
     assert np.all(
-        reference == parallel_funcs.threaded_take(make_test_array, idcs)
+        reference == parallel.threaded_take(make_test_array, idcs)
     ), "Failed n_threads>idcs.shape[0]"
 
-    parallel_funcs.set_num_threads(orig_num_threads)
+    parallel.set_num_threads(orig_num_threads)
 
 
 def test_take_axis(make_test_array):
@@ -52,26 +52,23 @@ def test_take_axis(make_test_array):
     idcs = np.random.randint(0, 100, size=(20, 3))
     reference = np.take(make_test_array, idcs, axis=1)
 
-    orig_num_threads = parallel_funcs.get_num_threads()
-    parallel_funcs.set_num_threads(1)
+    orig_num_threads = parallel.get_num_threads()
+    parallel.set_num_threads(1)
     assert np.all(
-        reference
-        == parallel_funcs.threaded_take(make_test_array, idcs, axis=1)
+        reference == parallel.threaded_take(make_test_array, idcs, axis=1)
     ), "Thin wrapper fail (failed at 1 thread)"
 
-    parallel_funcs.set_num_threads(2)
+    parallel.set_num_threads(2)
     assert np.all(
-        reference
-        == parallel_funcs.threaded_take(make_test_array, idcs, axis=1)
+        reference == parallel.threaded_take(make_test_array, idcs, axis=1)
     ), "Failed on 2 threads"
 
-    parallel_funcs.set_num_threads(21)
+    parallel.set_num_threads(21)
     assert np.all(
-        reference
-        == parallel_funcs.threaded_take(make_test_array, idcs, axis=1)
+        reference == parallel.threaded_take(make_test_array, idcs, axis=1)
     ), "Failed n_threads>idcs.shape[0]"
 
-    parallel_funcs.set_num_threads(orig_num_threads)
+    parallel.set_num_threads(orig_num_threads)
 
 
 def test_clip(make_test_array):
@@ -79,23 +76,23 @@ def test_clip(make_test_array):
     amin, amax = 0.25, 0.75
     reference = np.clip(make_test_array, amin, amax)
 
-    orig_num_threads = parallel_funcs.get_num_threads()
-    parallel_funcs.set_num_threads(1)
+    orig_num_threads = parallel.get_num_threads()
+    parallel.set_num_threads(1)
     assert np.all(
-        reference == parallel_funcs.threaded_clip(make_test_array, amin, amax)
+        reference == parallel.threaded_clip(make_test_array, amin, amax)
     ), "Thin wrapper fail (failed at 1 thread)"
 
-    parallel_funcs.set_num_threads(2)
+    parallel.set_num_threads(2)
     assert np.all(
-        reference == parallel_funcs.threaded_clip(make_test_array, amin, amax)
+        reference == parallel.threaded_clip(make_test_array, amin, amax)
     ), "Failed on 2 threads"
 
-    parallel_funcs.set_num_threads(101)
+    parallel.set_num_threads(101)
     assert np.all(
-        reference == parallel_funcs.threaded_clip(make_test_array, amin, amax)
+        reference == parallel.threaded_clip(make_test_array, amin, amax)
     ), "Failed n_threads>idcs.shape[0]"
 
-    parallel_funcs.set_num_threads(orig_num_threads)
+    parallel.set_num_threads(orig_num_threads)
 
 
 def test_elliptics(make_test_array):
@@ -104,21 +101,21 @@ def test_elliptics(make_test_array):
     refe = ellipe(make_test_array)
     refk = ellipk(make_test_array)
 
-    orig_num_threads = parallel_funcs.get_num_threads()
-    parallel_funcs.set_num_threads(1)
-    rese, resk = parallel_funcs.threaded_elliptics_ek(make_test_array)
+    orig_num_threads = parallel.get_num_threads()
+    parallel.set_num_threads(1)
+    rese, resk = parallel.threaded_elliptics_ek(make_test_array)
     assert np.all(
         (refe == rese) & (refk == resk)
     ), "Thin wrapper fail (failed at 1 thread)"
 
-    parallel_funcs.set_num_threads(2)
-    rese, resk = parallel_funcs.threaded_elliptics_ek(make_test_array)
+    parallel.set_num_threads(2)
+    rese, resk = parallel.threaded_elliptics_ek(make_test_array)
     assert np.all((refe == rese) & (refk == resk)), "Failed on 2 threads"
 
-    parallel_funcs.set_num_threads(101)
-    rese, resk = parallel_funcs.threaded_elliptics_ek(make_test_array)
+    parallel.set_num_threads(101)
+    rese, resk = parallel.threaded_elliptics_ek(make_test_array)
     assert np.all(
         (refe == rese) & (refk == resk)
     ), "Failed n_threads>idcs.shape[0]"
 
-    parallel_funcs.set_num_threads(orig_num_threads)
+    parallel.set_num_threads(orig_num_threads)
